pip_requirements.txt
absl-py==0.7.0
astor==0.7.1
astroid==2.0.4
certifi==2018.11.29
cloudpickle==0.8.0
colorama==0.3.9
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.16.1
Pillow==5.4.1
protobuf==3.6.1
pylint==2.1.1
pyparsing==2.3.1
python-dateutil==2.8.0
PyWavelets==1.0.1
PyYAML==3.13
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
toolz==0.9.0
typed-ast==1.1.0
Werkzeug==0.14.1
wincertstore==0.2
